Added the logic to generate users from providers
Changed the user to not belongs_to employee, but now the user has a
polymorphic association called authenticable.

Created the logic to generate users from provider, the ones which no
have any user.

Changed the way to add provider to price collection, need to fill email
and login to create the users.

Changed the logic to generate the proposal items, just update the
created proposals and add non existents items.

[#29939559]