Remove the stupid message telling user to login when he access the signin route
This was happening because the root route was pointing to
bookmarks#show and on first state of session, the user is not
authenticated.

Using this way, the root route will be bookmarks#show for authenticated
users and devise/sessions#new to unauthenticated users.

The after sign out path for was removed from application controller
because it's not necessary anymore

[finishes #35970763]