Add trading item classification
When there is only one bidder to the trading
user should be redirected to the classification
page.

Classification page should have all bidders
with proposals and their respective lowest 
proposal for the current item.

Should have too a percent based on the first
place of the classification, and should have
the position into the classification of each
bidder.

The percent value follow the rule:

(current_amount - amount_of_first) / 
amount_of_first * 100

[Finishes #39409397]