Add attr_header
The attr_header allows to create columns with 
headers to the index view automatically.

When the attr_header is defined the index renders
a different partial to list the collection of
resources, where it have columns with headers
defined at attr_header.

When the attd_header is not defined the index will
follow the default style like it was before this
change, rendering the default partial to list the
collection of resources.

I needed to override the Decore::Infection making
it smart to know when is decorating a class or a
instance.

This new feature is used at decorators including
Decore::Header and than using attr_header at the
decorator.

[Finishes #40012247]