Set PurchaseSolicitation when use the ItemGroup
When an item_group is used all it purchase 
solicitations should update their status to
'in_purchase_process'. And when this item_group
is exchanged or is not used anymore, than the
old item_group should change all it purchase
solicitations status to 'liberated' and the new
one should update all it purchase solicitations
status to 'in_purchase_process'.

[Finishes #39696181]