Fill in responsible when fill in PurchaseSolicitation
When on administrative process form and the purchase
solicitation is filled, then the responsible
should be filled in automatically from the
purchase solicitation. But the responsible is
filled in automatically only when it is empty.

Although the responsible is being filled in
automatically it can be changed by the user.

[Finishes #40561607]