Fix relationship between items and administrative processes and direct purchase.
The most important thing in this commit, is the
relationship between the purchase solicitation
budget allocation items and direct purchase and
administrative process.

Before, the attributes where changed based in the
status of themselves.
Now, the changes are based in the relationship.

Item group now, has one direct purchase and one
administrative process relationship.

Item has a new status value, partially_fulfilled.
Because we need differ when a item has a relationship
with a direct purchase with supply authorization, from
another direct purchase without supply authorization.

[Finishes #42285409]