Change rule for trading item negotiation
Remove auto-select bidder for negotiation. Now
the negotiation should be done bidder by bidder.

The button for negotiation appear only for the
bidders who can negotiate.

Now when there is not preemptive right bidders and
the winner can negotiate. If it negotiation fail
than the trading item cannot have a winner. If it
gave a better offer then the it will be the
winner.

Add rule to cannot close a trading_item with
status winner if the item has not a winner.

[Finishes #43642185]