Changed order of some models visualizations
Some models were ordering by id and that wasn't 
making much sense to the user. They have been ordered
by name or another alphabetical attribute.

It was, also, created an option to the orderize method.
Now it accepts a :on parameter that expects a association
name, so the order can be made by joining tables.

[Finishes #44640159]